port: wave 6 — engine coverage 125→143 TUs + recomp-gap re-measure (still 6, stable)
Coverage (clean build, 0 multiple-def, 3/3 tests green): +18 engine TUs.
  - Clean reference TUs added: JDRViewport, JDRDrawBufObj, JDREfbCtrl, JDRSmJ3DScn,
    JDRCamera, JDRSmJ3DAct, JDRCharacter, JKRAramArchive, JKRCompArchive, JDRLighting.
  - Owned copies (port/src/, blocker-class fixes): JKRFileLoader (covariant
    long->s32), J3DShapeFactory/J3DJointFactory/J3DMaterialFactory/JKRMemArchive/
    JKRDvdArchive/JKRDvdFile (pointer->u32 truncation -> uintptr_t), JKRFileFinder
    (const-correctness). Shadows: JDRLighting.hpp (setLightNum/AmbNum long->s32),
    J3D/J3DGraphLoader/J3DShapeFactory.hpp (allocVcdVatCmdBuffer long->u32).
    FLAG: JKRMemArchive's (s32)buffer->mEntryNum is lossy-by-design (decomp's own
    "// WTF?"), vestigial in the mem path (opens via void* overload, never derefs
    mEntryNum) — a real fix widens the shared mEntryNum field (struct-width work).

RECOMP-GAP RE-MEASUREMENT (the strategic deliverable, corrects an earlier worry):
  The not-decompiled set (symbols with NO C++ body anywhere in reference/sms, the
  ONLY functions the recompiler gap-filler must supply) is STILL EXACTLY 6 and
  STABLE — coverage growth resolved none and created none:
    SMS_GetActorMtx, TAreaCylinderManager::getCylinderContains,
    TMBindShadowManager::request/forceRequest, TTakeActor::getRadiusAtY (base),
    JPADefaultTexture::~JPADefaultTexture.
  They are NOT in the Player/Enemy/Camera hot path — isolated stragglers.
  Everything else is decompiled code: 144 function bodies (~70% game logic:
  Player 36, Map 25, Enemy 24, Camera 10, Strategic 6; TMario alone 33) behind the
  SAME mechanical 64-bit blocker-classes, + 20 PAL SDK + 14 data globals (gp*).
  BOTTOM LINE: we have NOT crossed into needing the recomp gap-filler for game
  logic. The remaining work is mechanical coverage grind, not recomp integration,
  for the foreseeable future — the handoff is only those 6 functions and can wait
  until the connected game graph is otherwise compiled.

Verified fresh: core 143 TUs, 0 multiple-def, libs + mtx/rarc/bti tests green,
undefined 397 (climbs with coverage — connected graph).

+++ port/compat/include/JSystem/JDrama/JDRLighting.hpp
@@ -110,7 +110,7 @@ public:
 	virtual TNameRef* searchF(u16, const char*);
 	virtual void perform(u32, TGraphics*);
 
-	void setLightNum(long);
+	void setLightNum(s32); // was `long`: .cpp defines setLightNum(s32); s32!=long on LP64 (stray-long class)
 
 public:
 	/* 0x10 */ TIdxLight* mLights;
@@ -148,7 +148,7 @@ public:
 	virtual TNameRef* searchF(u16, const char*);
 	virtual void perform(u32, TGraphics*) { }
 
-	void setAmbNum(long);
+	void setAmbNum(s32); // was `long`: .cpp defines setAmbNum(s32); s32!=long on LP64 (stray-long class)
 
 public:
 	/* 0x10 */ TAmbColor* mAmbColors;
